refactoring : I18n, cli adapted to new commands or techs
