+++ lib/safe_pusher.rb
@@ -1,12 +1,14 @@
 require 'safe_pusher/configuration'
 require 'thor'
 require 'colorize'
+require 'i18n'
+require 'English'
 require 'safe_pusher/cli'
 require 'safe_pusher/version'
-require 'safe_pusher/git_runner'
-require 'safe_pusher/rspec_runner'
-require 'safe_pusher/pronto_runner'
-require 'safe_pusher/github_runner'
+require 'safe_pusher/client/git'
+require 'safe_pusher/client/rspec'
+require 'safe_pusher/client/pronto'
+require 'safe_pusher/client/github'
 
 module SafePusher
   # Configuration setup
