+++ safe_pusher.gemspec
@@ -37,6 +37,7 @@ Gem::Specification.new do |spec|
   spec.require_paths = ['lib']
 
   spec.add_dependency 'colorize', '~> 0.8.1'
+  spec.add_dependency 'i18n', '~> 1.6.0'
 
   spec.add_development_dependency 'bundler', '>= 1.16'
   spec.add_development_dependency 'fashion_police', '~> 1.2'
